requirements.txt
boto3
GitPython